feat(ui): pipeline
add blocks for formats
