+++ app/pages/7_📚Documentation du traitement batch.py
